requirements.txt
streamlit>=1.28.0
geopy>=2.4.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
geopandas>=0.14.0